static renderer wip
