+++ changelog.md
@@ -1,0 +1,3 @@
+# Changelog from 2023.10.2
+
+- remove network status from node-status-context
