+++ frontend/apps/site/publication-page.tsx
@@ -1,33 +1,22 @@
 import {
   Account,
-  Block,
-  EmbedBlock,
-  FileBlock,
-  HMBlockChildrenType,
-  HeadingBlock,
-  ImageBlock,
-  InlineContent,
-  ParagraphBlock,
-  PresentationBlock,
+  HMBlockFile,
+  HMBlockNode,
+  HMBlockVideo,
+  HMGroup,
+  HMPublication,
   Publication,
   UnpackedHypermediaId,
-  VideoBlock,
   createHmDocLink,
-  createPublicWebHmUrl,
   formatBytes,
   getCIDFromIPFSUrl,
   groupDocUrl,
-  idToUrl,
-  isHypermediaScheme,
-  toHMInlineContent,
-  unpackDocId,
   unpackHmId,
+  StaticBlockNode,
 } from '@mintter/shared'
 import {
   ArrowRight,
   Button,
-  Copy,
-  ExternalLink,
   File,
   FontSizeTokens,
   PageSection,
@@ -35,23 +24,18 @@ import {
   SideSection,
   SideSectionTitle,
   SizableText,
-  Spinner,
   Tooltip,
   XStack,
   YStack,
   useMedia,
 } from '@mintter/ui'
 import {DehydratedState} from '@tanstack/react-query'
-import {cidURL} from 'ipfs'
 import {NextLink} from 'next-link'
 import Head from 'next/head'
-import {useRouter} from 'next/router'
-import {useMemo, useState} from 'react'
-import {DimensionValue} from 'react-native'
+import {useMemo} from 'react'
 import {WebTipping} from 'web-tipping'
 import Footer from './footer'
 import {PublicationMetadata} from './publication-metadata'
-import {HMBlock, HMBlockNode, HMGroup, HMPublication} from './server/json-hm'
 import {SiteHead} from './site-head'
 import {trpc} from './trpc'
 
@@ -72,7 +56,7 @@ export type PublicationPageData = {
 
 let blockVerticalPadding: FontSizeTokens = '$4'
 let blockHorizontalPadding: FontSizeTokens = '$4'
-let blockBorderRadius: FontSizeTokens = '$3'
+let blockBorderRadius = '$3'
 
 function OpenInAppLink({url}: {url: string}) {
   return (
@@ -115,11 +99,7 @@ export default function PublicationPage({
         <meta name="hypermedia-entity-version" content={pub?.version} />
         <meta name="hypermedia-entity-title" content={pub?.document?.title} />
       </Head>
-      <SiteHead
-        siteTitle={contextGroup?.title}
-        pageTitle={pub?.document?.title}
-        siteSubheading={contextGroup?.description}
-      />
+      <SiteHead pageTitle={pub?.document?.title} />
       <PageSection.Root>
         <PublicationContextSidebar
           group={contextGroup}
@@ -172,423 +152,6 @@ export default function PublicationPage({
       </PageSection.Root>
       <Footer />
     </>
-  )
-}
-
-function InlineContentView({
-  inline,
-  style,
-}: {
-  inline: InlineContent[]
-  style?: {heading: boolean}
-}) {
-  return (
-    <>
-      {inline.map((content, index) => {
-        if (content.type === 'text') {
-          let textDecorationLine:
-            | 'underline'
-            | 'none'
-            | 'line-through'
-            | 'underline line-through'
-            | '' = ''
-          if (content.styles.underline) {
-            if (content.styles.strike) {
-              textDecorationLine = 'underline line-through'
-            } else {
-              textDecorationLine = 'underline'
-            }
-          } else if (content.styles.strike) {
-            textDecorationLine = 'line-through'
-          }
-
-          let children: any = content.text
-
-          if (content.styles.bold) {
-            children = <b>{children}</b>
-          }
-
-          if (content.styles.italic) {
-            children = <i>{children}</i>
-          }
-
-          if (content.styles.underline) {
-            children = <u>{children}</u>
-          }
-
-          if (content.styles.code) {
-            children = (
-              <SizableText
-                paddingHorizontal="$2"
-                paddingVertical="$1"
-                backgroundColor="$color7"
-                borderRadius="$2"
-              >
-                <code style={{fontSize: '0.95em'}}>{children}</code>
-              </SizableText>
-            )
-          }
-
-          if (content.styles.backgroundColor) {
-            children = (
-              <span style={{backgroundColor: content.styles.backgroundColor}}>
-                {children}
-              </span>
-            )
-          }
-
-          if (content.styles.strike) {
-            children = <s>{children}</s>
-          }
-
-          if (content.styles.textColor) {
-            children = (
-              <SizableText color={content.styles.textColor}>
-                {children}
-              </SizableText>
-            )
-          }
-
-          return (
-            <SizableText
-              size="$6"
-              key={index}
-              textDecorationLine={textDecorationLine || undefined}
-            >
-              {children}
-            </SizableText>
-          )
-        }
-        if (content.type === 'link') {
-          const href = isHypermediaScheme(content.href)
-            ? idToUrl(content.href, null)
-            : content.href
-
-          const isExternal = isHypermediaScheme(content.href)
-          return (
-            href && (
-              <Tooltip content={href}>
-                <a
-                  href={href}
-                  key={index}
-                  className={isExternal ? 'hm-link' : 'link'}
-                  style={{
-                    cursor: 'pointer',
-                    display: 'inline-block',
-                  }}
-                >
-                  <InlineContentView inline={content.content} />
-                  {isExternal ? <ExternalLink size={10} /> : null}
-                </a>
-              </Tooltip>
-            )
-          )
-        }
-        return null
-      })}
-    </>
-  )
-}
-
-function StaticSectionBlock({block}: {block: HeadingBlock | ParagraphBlock}) {
-  const inline = useMemo(() => toHMInlineContent(new Block(block)), [block])
-  const isHeading = block.type == 'heading'
-  return (
-    <YStack
-      id={`${block.id}-block`}
-      // paddingLeft={isBlockquote ? 20 : 0}
-      // borderLeftWidth={isBlockquote ? 1 : 0}
-      borderLeftColor={'blue'}
-    >
-      <SizableText
-        size={isHeading ? '$7' : undefined}
-        fontWeight={isHeading ? 'bold' : undefined}
-        tag={isHeading ? 'h2' : 'p'}
-      >
-        <InlineContentView
-          inline={inline}
-          style={{
-            heading: block.type === 'heading',
-          }}
-        />
-      </SizableText>
-    </YStack>
-  )
-}
-
-function StaticImageBlock({block}: {block: ImageBlock}) {
-  const cid = getCIDFromIPFSUrl(block?.ref)
-  if (!cid) return null
-  return (
-    <XStack
-      backgroundColor="$color3"
-      borderColor="$color4"
-      borderWidth={1}
-      borderRadius={blockBorderRadius}
-      overflow="hidden"
-      hoverStyle={{
-        backgroundColor: '$color4',
-      }}
-    >
-      <img
-        id={`${block.id}-block`}
-        alt={block.attributes?.alt}
-        src={cidURL(cid)}
-        className="image"
-        onError={(e) => {
-          console.error('image errored', e)
-        }}
-      />
-    </XStack>
-  )
-  // return <img src={`${process.env.GRPC_HOST}/ipfs/${cid}`} />
-}
-
-function stripHMLinkPrefix(link: string) {
-  return link.replace(/^hm:\//, '')
-}
-
-function StaticEmbedBlock({block}: {block: EmbedBlock}) {
-  const reference = block.ref
-  const docId = unpackDocId(reference)
-  const router = useRouter()
-  let embed = trpc.publication.get.useQuery(
-    {
-      documentId: docId?.docId,
-      versionId: docId?.version || undefined,
-    },
-    {enabled: !!docId},
-  )
-  let content = <Spinner />
-  if (embed.data?.publication?.document?.children) {
-    if (docId?.blockRef) {
-      const blockNode = getBlockNodeById(
-        embed.data?.publication?.document?.children,
-        docId.blockRef,
-      )
-      content = blockNode ? (
-        <StaticBlockNode block={blockNode} />
-      ) : (
-        <SizableText>Block not found.</SizableText>
-      )
-    } else {
-      content = (
-        <>
-          {embed.data?.publication?.document?.children?.map((block) => (
-            <StaticBlockNode block={block} key={block?.block?.id} ctx={{}} />
-          ))}
-        </>
-      )
-    }
-  }
-  return (
-    <YStack
-      id={`${block.id}-block`}
-      data-ref={reference}
-      transform="translateX(-19px)"
-      width="calc(100% + 16px)"
-      position="relative"
-    >
-      <YStack
-        // padding={blockVerticalPadding}
-        // paddingVertical={blockVerticalPadding}
-        // borderRadius={blockBorderRadius}
-        // backgroundColor="$color5"
-        // marginRight={blockHorizontalPadding}
-        hoverStyle={{
-          cursor: 'pointer',
-        }}
-        onPress={() => {
-          const ref = stripHMLinkPrefix(block.ref)
-          router.push(ref)
-        }}
-        href={stripHMLinkPrefix(block.ref)}
-      >
-        {content}
-        {/* <EmbedMetadata embed={embed} /> */}
-      </YStack>
-    </YStack>
-  )
-}
-
-function StaticBlock({block, isList}: {block: HMBlock; isList?: boolean}) {
-  let niceBlock = block as PresentationBlock // todo, validation
-
-  if (niceBlock.type == 'paragraph' || niceBlock.type == 'heading') {
-    return <StaticSectionBlock block={niceBlock} isList={isList} />
-  }
-
-  if (niceBlock.type == 'image') {
-    return <StaticImageBlock block={niceBlock} isList={isList} />
-  }
-  if (niceBlock.type == 'embed') {
-    return <StaticEmbedBlock block={niceBlock} isList={isList} />
-  }
-  if (niceBlock.type == 'code') {
-    let _content = (
-      <pre>
-        <code>{block.text}</code>
-      </pre>
-    )
-
-    return isList ? <li>{_content}</li> : _content
-  }
-
-  if (niceBlock.type == 'file') {
-    return <StaticFileBlock block={niceBlock} />
-  }
-
-  if (niceBlock.type == 'video') {
-    return <StaticVideoBlock block={niceBlock} isList={isList} />
-  }
-  // fallback for unknown block types
-  // return <span>{JSON.stringify(block)}</span>
-  return (
-    <ErrorMessageBlock
-      // @ts-expect-error
-      id={`${niceBlock.id}-block`}
-      // @ts-expect-error
-      message={`Unknown block type: "${niceBlock.type}"`}
-    />
-  )
-}
-
-function ErrorMessageBlock({message, id}: {message: string; id: string}) {
-  return (
-    <YStack
-      padding={blockVerticalPadding}
-      paddingVertical={blockVerticalPadding}
-      marginRight={blockHorizontalPadding}
-      backgroundColor="$color3"
-      borderColor="$color4"
-      borderWidth={1}
-      borderRadius={blockBorderRadius}
-      overflow="hidden"
-      hoverStyle={{
-        backgroundColor: '$color4',
-      }}
-      id={id}
-    >
-      <SizableText color="$color8" size="$6">
-        {message}
-      </SizableText>
-    </YStack>
-  )
-}
-
-type PublicationViewContext = {
-  headingDepth?: number
-  enableBlockCopy?: boolean
-  ref?: {
-    docId?: string
-    docVersion?: string
-  }
-}
-
-function StaticBlockNode({
-  childrenType,
-  block,
-  ctx,
-}: {
-  block: HMBlockNode
-  ctx?: PublicationViewContext
-  childrenType?: HMBlockChildrenType
-}) {
-  const [isHovering, setIsHovering] = useState(false)
-
-  const children =
-    (block.children?.length || 0) > 0 ? (
-      <YStack
-        onHoverIn={() => setIsHovering(false)}
-        onHoverOut={() => setIsHovering(true)}
-        tag={block.block?.attributes?.childrenType}
-        start={
-          block.block?.attributes?.childrenType == 'ol'
-            ? (block.block?.attributes?.start as DimensionValue) || 1
-            : 'auto'
-        }
-        display="block"
-      >
-        {block.children?.map((child, index) => (
-          <StaticBlockNode
-            childrenType={
-              block.block?.attributes?.childrenType as HMBlockChildrenType
-            }
-            key={child.block?.id || index}
-            block={child}
-            ctx={ctx}
-          />
-        ))}
-      </YStack>
-    ) : null
-  const id = block.block?.id || 'unknown-block'
-  const isList = childrenType == 'ol' || childrenType == 'ul'
-  const docIds = ctx?.ref?.docId ? unpackDocId(ctx?.ref?.docId) : null
-  return (
-    <YStack
-      paddingVertical="$2"
-      paddingLeft={blockHorizontalPadding}
-      id={id}
-      onHoverIn={() => setIsHovering(true)}
-      onHoverOut={() => setIsHovering(false)}
-      position="relative"
-      // borderWidth={1}
-      // borderColor="$color6"
-      borderRadius={blockBorderRadius}
-      backgroundColor={isHovering ? '$backgroundHover' : 'transparent'}
-      // backgroundColor={
-      //   window.location.hash == `#${block.block?.id}`
-      //     ? '$yellow5'
-      //     : isHovering
-      //     ? '$color4'
-      //     : 'transparent'
-      // }
-      tag={isList ? 'li' : undefined}
-      marginLeft={isList ? blockHorizontalPadding : undefined}
-      style={
-        isList
-          ? {
-              display: 'list-item',
-              listStyleType: childrenType == 'ol' ? 'decimal' : 'disc',
-            }
-          : undefined
-      }
-    >
-      {block.block && <StaticBlock isList={isList} block={block.block} />}
-      {children}
-      {ctx?.enableBlockCopy && (
-        <XStack
-          bg="$background"
-          position="absolute"
-          top={0}
-          right={0}
-          opacity={isHovering ? 1 : 0}
-          overflow="hidden"
-          borderRadius={blockBorderRadius}
-        >
-          <Tooltip content={`Copy block reference`}>
-            {docIds?.eid && (
-              <Button
-                tag="a"
-                size="$2"
-                chromeless
-                href={`#${id}`}
-                icon={Copy}
-                onPress={() => {
-                  navigator.clipboard.writeText(
-                    createPublicWebHmUrl('d', docIds?.eid, {
-                      version: ctx?.ref?.docVersion,
-                      blockRef: id,
-                      hostname: `${window.location.protocol}//${window.location.host}`,
-                    }),
-                  )
-                }}
-              />
-            )}
-          </Tooltip>
-        </XStack>
-      )}
-    </YStack>
   )
 }
 
@@ -655,36 +218,13 @@ export function PublicationContent({
             enableBlockCopy: true,
             ref: {
               docId: publication?.document?.id,
-              docVersion: publication.version,
+              docVersion: publication?.version,
             },
           }}
         />
       ))}
     </YStack>
   )
-}
-
-function getBlockNodeById(
-  blocks: Array<HMBlockNode>,
-  blockId: string,
-): HMBlockNode | null {
-  if (!blockId) return null
-
-  let res: HMBlockNode | undefined
-  blocks.find((bn) => {
-    if (bn.block?.id == blockId) {
-      res = bn
-      return true
-    } else if (bn.children?.length) {
-      const foundChild = getBlockNodeById(bn.children, blockId)
-      if (foundChild) {
-        res = foundChild
-        return true
-      }
-    }
-    return false
-  })
-  return res || null
 }
 
 export function useGroupContentUrl(
@@ -766,6 +306,7 @@ function GroupSidebarContent({
         if (!item || !groupId?.eid) return null
         return (
           <GroupSidebarContentItem
+            key={item.docId}
             item={item}
             groupId={groupId}
             groupVersion={group?.version}
@@ -780,10 +321,12 @@ function GroupSidebarContent({
 function PublicationContextSidebar({
   group,
   activePathName,
+  display,
   ...props
 }: {
   group?: HMGroup | null
   activePathName: string
+  display: any
 }) {
   const groupContent = trpc.group.listContent.useQuery(
     {
@@ -800,14 +343,12 @@ function PublicationContextSidebar({
     />
   ) : null
 
-  return (
-    <PageSection.Side show={props.display != 'none'} {...props}>
-      {groupSidebarContent}
-    </PageSection.Side>
-  )
+  return display != 'none' ? (
+    <PageSection.Side {...props}>{groupSidebarContent}</PageSection.Side>
+  ) : null
 }
 
-export function StaticFileBlock({block}: {block: FileBlock}) {
+export function StaticFileBlock({block}: {block: HMBlockFile}) {
   let cid = useMemo(() => getCIDFromIPFSUrl(block.ref), [block.ref])
   return (
     <NextLink
@@ -820,7 +361,7 @@ export function StaticFileBlock({block}: {block: FileBlock}) {
           backgroundColor="$color3"
           borderColor="$color4"
           borderWidth={1}
-          borderRadius={blockBorderRadius}
+          borderRadius={blockBorderRadius as any}
           // marginRight={blockHorizontalPadding}
           overflow="hidden"
           hoverStyle={{
@@ -867,7 +408,7 @@ function StaticVideoBlock({
   block,
   isList,
 }: {
-  block: VideoBlock
+  block: HMBlockVideo
   isList?: boolean
 }) {
   const ref = block.ref || ''
