+++ frontend/packages/app/pages/blocknote.tsx
@@ -1,6 +1,5 @@
 import {
   BlockNoteEditor,
-  DefaultBlockSchema,
   defaultBlockSchema,
   HMBlockSchema,
 } from '@mintter/editor'
