+++ frontend/packages/app/pages/group.tsx
@@ -1,9 +1,10 @@
 import Footer from '@mintter/app/components/footer'
-import {StaticBlockNode} from '@mintter/editor'
 import {
   Document,
   Group,
+  HMBlockNode,
   Role,
+  StaticBlockNode,
   formattedDate,
   idToUrl,
   pluralS,
@@ -35,9 +36,12 @@ import {
   Store,
   Trash,
 } from '@tamagui/lucide-icons'
+import {Allotment} from 'allotment'
+import 'allotment/dist/style.css'
 import {useEffect, useMemo, useRef, useState} from 'react'
 import {toast} from 'react-hot-toast'
 import {AccountLinkAvatar} from '../components/account-link-avatar'
+import {EntityVersionsAccessory} from '../components/changes-list'
 import {useAppDialog} from '../components/dialog'
 import {useEditGroupInfoDialog} from '../components/edit-group-info'
 import {FooterButton} from '../components/footer'
@@ -45,6 +49,7 @@ import {AppLinkText} from '../components/link'
 import {OptionsDropdown, copyLinkMenuItem} from '../components/list-item'
 import {PublicationListItem} from '../components/publication-list-item'
 import {EditDocActions} from '../components/titlebar/common'
+import {VersionChangesInfo} from '../components/version-changes-info'
 import {useAccount, useMyAccount} from '../models/accounts'
 import {useAllChanges} from '../models/changes'
 import {useDraftList, usePublication} from '../models/documents'
@@ -56,16 +61,12 @@ import {
   useRemoveDocFromGroup,
   useRenameGroupDoc,
 } from '../models/groups'
+import {useOpenUrl} from '../open-url'
 import {GroupRoute, useNavRoute} from '../utils/navigation'
 import {useOpenDraft} from '../utils/open-draft'
 import {pathNameify} from '../utils/path'
-import {useNavigate} from '../utils/useNavigate'
-import {Allotment} from 'allotment'
-import 'allotment/dist/style.css'
-import {useOpenUrl} from '../open-url'
-import {EntityVersionsAccessory} from '../components/changes-list'
-import {VersionChangesInfo} from '../components/version-changes-info'
 import {hostnameStripProtocol} from '../utils/site-hostname'
+import {useNavigate} from '../utils/useNavigate'
 
 export function RenamePubDialog({
   input: {groupId, pathName, docTitle},
