+++ frontend/packages/shared/src/index.ts
@@ -3,3 +3,5 @@ export * from './constants'
 export * from './hm-documents'
 export * from './grpc-client'
 export * from './utils'
+export * from './json-hm'
+export * from './static-renderer'
