+++ frontend/packages/shared/src/json-hm.ts
@@ -1,8 +1,5 @@
 import type {
   Account,
-  Annotation,
-  Block,
-  BlockNode,
   ChangeInfo,
   Device,
   Document,
@@ -23,50 +20,8 @@ export type HMChangeInfo = {
   deps?: string[]
 }
 
-export type ServerAnnotation = Annotation
-export type HMAnnotation = {
-  type?: string
-  ref?: string
-  attributes?: {[key: string]: string}
-  starts?: number[]
-  ends?: number[]
-}
-
-export type ServerBlock = Block
-export type HMBlock = {
-  id?: string
-  type?: string
-  text?: string
-  ref?: string
-  attributes?: {[key: string]: string}
-  annotations?: HMAnnotation[]
-  revision?: string
-}
-
-export type ServerBlockNode = BlockNode
-export type HMBlockNode = {
-  block?: HMBlock
-  children?: HMBlockNode[]
-}
-
 export type ServerDocument = Document
-export type HMDocument = {
-  title?: string
-  id?: string
-  author?: string
-  webUrl?: string
-  editors?: string[]
-  children?: HMBlockNode[]
-  createTime?: HMTimestamp
-  updateTime?: HMTimestamp
-  publishTime?: HMTimestamp
-}
-
 export type ServerPublication = Publication
-export type HMPublication = {
-  document?: HMDocument
-  version?: string
-}
 
 export type ServerGroupSiteInfo = Group_SiteInfo
 export type HMGroupSiteInfo = {
