+++ frontend/packages/ui/tsconfig.json
@@ -3,7 +3,7 @@
   "include": ["src"],
   "compilerOptions": {
     "composite": true,
-    "jsx": "react-jsx",
+    "jsx": "react-jsx"
   },
-  "references": [],
+  "references": []
 }
