+++ issues.md
@@ -1,0 +1,8 @@
+# issues
+
+- [ ] `getPeerInfo` is returning undefined when offline
+- [ ] make sure the correct production version is passed to the app build in CI. currently the version is 0.0.0 in the production app (2023.10.2)
+- [ ] document models
+  - [ ] remove backgroundcolor props from editor model
+  - [ ] remove color from editor model
+  - [ ]
